+++ src/__tests__/config.test.ts
@@ -1,3 +1,5 @@
+import { makeConfig } from '../config';
+
 const defaultConfig: Record<string, string> = {
   command: 'up',
   'stack-name': 'dev',
@@ -5,29 +7,30 @@ const defaultConfig: Record<string, string> = {
   'cloud-url': 'file://~',
   'github-token': 'n/a',
   'pulumi-version': '^3',
+  'comment-on-pr': 'false',
+  upsert: 'false',
+  remove: 'false',
+  refresh: 'false',
+  'edit-pr-comment': 'false',
+  'expect-no-changes': 'false',
+  diff: 'false',
+  'target-dependents': 'false',
 };
+
+function setupMockedConfig(config: Record<string, string>) {
+  Object.entries(config).forEach(([key, value]) => {
+    process.env[`INPUT_${key.replace(/ /g, '_').toUpperCase()}`] = value;
+  });
+}
 
 describe('config.ts', () => {
   beforeEach(() => {
     jest.resetModules();
   });
   it('should validate a configuration', async () => {
-    const config = {
-      ...defaultConfig,
-      'comment-on-pr': 'false',
-    };
-    jest.mock('@actions/core', () => ({
-      getInput: jest.fn((name: string) => {
-        return config[name];
-      }),
-    }));
-    jest.mock('@actions/github', () => ({
-      context: {},
-    }));
-
-    const { makeConfig } = require('../config');
-
+    setupMockedConfig(defaultConfig);
     const c = await makeConfig();
+
     expect(c).toBeTruthy();
     expect(c).toMatchInlineSnapshot(`
       Object {
@@ -36,152 +39,40 @@ describe('config.ts', () => {
         "commentOnPr": false,
         "commentOnPrNumber": undefined,
         "configMap": undefined,
+        "editCommentOnPr": false,
         "githubToken": "n/a",
-        "isPullRequest": false,
         "options": Object {
           "color": undefined,
-          "diff": undefined,
-          "editCommentOnPr": undefined,
-          "expectNoChanges": undefined,
-          "message": undefined,
+          "diff": false,
+          "expectNoChanges": false,
+          "message": "",
           "parallel": undefined,
-          "policyPackConfigs": undefined,
-          "policyPacks": undefined,
-          "pulumiVersion": "^3",
-          "replace": undefined,
-          "target": undefined,
-          "targetDependents": undefined,
-          "userAgent": "pulumi/actions@v4",
+          "policyPackConfigs": Array [],
+          "policyPacks": Array [],
+          "replace": Array [],
+          "target": Array [],
+          "targetDependents": false,
+          "userAgent": "pulumi/actions@v3",
         },
-        "refresh": undefined,
-        "remove": undefined,
-        "secretsProvider": undefined,
+        "pulumiVersion": "^3",
+        "refresh": false,
+        "remove": false,
+        "secretsProvider": "",
         "stackName": "dev",
-        "upsert": undefined,
+        "upsert": false,
         "workDir": "./",
       }
     `);
   });
   it('should fail if configuration are invalid', async () => {
-    const config: Record<string, string> = {
-      command: 'sideways',
-    };
-    jest.mock('@actions/core', () => ({
-      getInput: jest.fn((name: string) => {
-        return config[name];
-      }),
-    }));
-    jest.mock('@actions/github', () => ({
-      context: {},
-    }));
-
-    const { makeConfig } = require('../config');
-
-    await expect(makeConfig()).rejects.toThrow();
-  });
-  it('should validate a configuration with commentOnPr eq true', async () => {
-    const config = {
+    setupMockedConfig({
       ...defaultConfig,
-      'comment-on-pr': 'true',
-    };
-    jest.mock('@actions/core', () => ({
-      getInput: jest.fn((name: string) => {
-        return config[name];
-      }),
-    }));
+      'stack-name': '',
+      command: 'invalid',
+    });
 
-    const { makeConfig } = require('../config');
-
-    const c = await makeConfig();
-    expect(c).toBeTruthy();
-    expect(c).toMatchInlineSnapshot(`
-      Object {
-        "cloudUrl": "file://~",
-        "command": "up",
-        "commentOnPr": true,
-        "commentOnPrNumber": undefined,
-        "configMap": undefined,
-        "githubToken": "n/a",
-        "isPullRequest": false,
-        "options": Object {
-          "color": undefined,
-          "diff": undefined,
-          "editCommentOnPr": undefined,
-          "expectNoChanges": undefined,
-          "message": undefined,
-          "parallel": undefined,
-          "policyPackConfigs": undefined,
-          "policyPacks": undefined,
-          "pulumiVersion": "^3",
-          "replace": undefined,
-          "target": undefined,
-          "targetDependents": undefined,
-          "userAgent": "pulumi/actions@v4",
-        },
-        "refresh": undefined,
-        "remove": undefined,
-        "secretsProvider": undefined,
-        "stackName": "dev",
-        "upsert": undefined,
-        "workDir": "./",
-      }
-    `);
-  });
-  it('should determine when in a PR', async () => {
-    const config = {
-      ...defaultConfig,
-      'comment-on-pr': 'false',
-    };
-    jest.mock('@actions/core', () => ({
-      getInput: jest.fn((name: string) => {
-        return config[name];
-      }),
-    }));
-    jest.mock('@actions/github', () => ({
-      context: {
-        payload: {
-          pull_request: {
-            number: 5678,
-          },
-        },
-      },
-    }));
-
-    const { makeConfig } = require('../config');
-
-    const c = await makeConfig();
-    expect(c).toBeTruthy();
-    expect(c).toMatchInlineSnapshot(`
-      Object {
-        "cloudUrl": "file://~",
-        "command": "up",
-        "commentOnPr": false,
-        "commentOnPrNumber": undefined,
-        "configMap": undefined,
-        "githubToken": "n/a",
-        "isPullRequest": true,
-        "options": Object {
-          "color": undefined,
-          "diff": undefined,
-          "editCommentOnPr": undefined,
-          "expectNoChanges": undefined,
-          "message": undefined,
-          "parallel": undefined,
-          "policyPackConfigs": undefined,
-          "policyPacks": undefined,
-          "pulumiVersion": "^3",
-          "replace": undefined,
-          "target": undefined,
-          "targetDependents": undefined,
-          "userAgent": "pulumi/actions@v4",
-        },
-        "refresh": undefined,
-        "remove": undefined,
-        "secretsProvider": undefined,
-        "stackName": "dev",
-        "upsert": undefined,
-        "workDir": "./",
-      }
-    `);
+    await expect(() => makeConfig()).toThrowErrorMatchingInlineSnapshot(
+      `"Input was not correct for command. Valid alternatives are: up, update, refresh, destroy, preview"`,
+    );
   });
 });